+++ .github/workflows/mutation.yml
@@ -26,7 +26,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        package: ["core", "diff", "docusaurus", "printer-legacy", "utils"]
+        package: ["core", "diff", "docusaurus", "printer-legacy", "utils", "helpers"]
 
     steps:
       - name: Checkout Code Base
