+++ package.json
@@ -190,6 +190,7 @@
     "@graphql-markdown/core": "file:packages/core",
     "@graphql-markdown/diff": "file:packages/diff",
     "@graphql-markdown/docusaurus": "file:packages/docusaurus",
+    "@graphql-markdown/helpers": "file:packages/helpers",
     "@graphql-markdown/printer-legacy": "file:packages/printer-legacy",
     "@graphql-markdown/tools-config": "file:config",
     "@graphql-markdown/types": "file:packages/types",
