+++ packages/docusaurus/tests/__data__/docusaurus2-graphql-doc-generator-groups.config.js
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
+const { getTypeDirectiveValues } = require("@graphql-markdown/utils");
 const {
-  getTypeDirectiveValues,
   directiveDescriptor,
   directiveTag,
-} = require("@graphql-markdown/utils");
+} = require("@graphql-markdown/helpers");
 
 module.exports = {
   id: "schema_with_grouping",
