+++ packages/utils/src/index.ts
@@ -7,5 +7,4 @@ export * from "./url";
 export * from "./graphql";
 export * from "./group";
 export * from "./directive";
-export * from "./helper";
 export * from "./logger";
